+++ app/src/main/java/ca/mcgill/a11y/image/PollingService.java
@@ -20,7 +20,6 @@ import android.app.Service;
 import android.content.Intent;
 import android.os.Handler;
 import android.os.IBinder;
-import android.util.Log;
 
 import androidx.annotation.Nullable;
