+++ app/src/main/java/ca/mcgill/a11y/image/selectors/PhotoSelector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 IMAGE Project, Shared Reality Lab, McGill University
+ * Copyright (c) 2024 IMAGE Project, Shared Reality Lab, McGill University
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -14,42 +14,29 @@
  * and our Additional Terms along with this program.
  * If not, see <https://github.com/Shared-Reality-Lab/IMAGE-Monarch/LICENSE>.
  */
-package ca.mcgill.a11y.image;
+package ca.mcgill.a11y.image.selectors;
 
-
-import static ca.mcgill.a11y.image.DataAndMethods.confirmButton;
-import static ca.mcgill.a11y.image.DataAndMethods.displayGraphic;
 import static ca.mcgill.a11y.image.DataAndMethods.fileSelected;
-import static ca.mcgill.a11y.image.DataAndMethods.image;
 import static ca.mcgill.a11y.image.DataAndMethods.keyMapping;
-import static ca.mcgill.a11y.image.DataAndMethods.speaker;
+import static ca.mcgill.a11y.image.DataAndMethods.pingsPlayer;
 import static ca.mcgill.a11y.image.DataAndMethods.update;
 
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.media.MediaPlayer;
-import android.os.BrailleDisplay;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.MotionEvent;
-import android.widget.EditText;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.view.GestureDetectorCompat;
-import androidx.lifecycle.Observer;
-
 import org.json.JSONException;
-import org.xml.sax.SAXException;
-
 import java.io.IOException;
+import ca.mcgill.a11y.image.BaseActivity;
+import ca.mcgill.a11y.image.renderers.BasicPhotoMapRenderer;
+import ca.mcgill.a11y.image.DataAndMethods;
+import ca.mcgill.a11y.image.R;
 
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.xpath.XPathExpressionException;
-
-// generates map request using latitude and longitude coordinates and allows for selecting among map renderer(s) 
-public class MapSelector extends AppCompatActivity implements MediaPlayer.OnCompletionListener{
-    private BrailleDisplay brailleServiceObj = null;
+// generates photo request by navigating files in specified directory and allows for selecting among photo renderer(s)
+public class PhotoSelector extends BaseActivity implements MediaPlayer.OnCompletionListener{
 
 
     @SuppressLint("WrongConstant")
@@ -57,40 +44,46 @@ public class MapSelector extends AppCompatActivity implements MediaPlayer.OnComp
     protected void onCreate(Bundle savedInstanceState) {
         Intent intent = getIntent();
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_map_selector);
+        //setContentView(R.layout.activity_photo_selector);
 
         //mDetector = new GestureDetectorCompat(getApplicationContext(),this);
         // Set the gesture detector as the double tap
         //mDetector.setOnDoubleTapListener(this);
 
-        brailleServiceObj = DataAndMethods.brailleServiceObj;
-        // DataAndMethods.initialize(brailleServiceObj, getApplicationContext(), findViewById(android.R.id.content));
-        DataAndMethods.update.observe(this,new Observer<Boolean>() {
-            @Override
-            public void onChanged(Boolean changedVal) {
-                if (changedVal){
-                    startActivity(new Intent(getApplicationContext(), BasicPhotoMapRenderer.class));
-                }
-            }
-
-        });
+        //brailleServiceObj = DataAndMethods.brailleServiceObj;
+        //DataAndMethods.initialize(brailleServiceObj, getApplicationContext(), findViewById(android.R.id.content));
     }
 
     @Override
     public boolean onKeyDown(int keyCode, KeyEvent event) {
-        super.onKeyDown(keyCode, event);
+        //super.onKeyDown(keyCode, event);
         switch (keyMapping.getOrDefault(keyCode, "default")) {
-            case "OK":
+            // Navigating between files
+            case "UP":
+                Log.d("KEY EVENT", event.toString());
                 try {
-                    Double latitude= Double.parseDouble(((EditText) findViewById(R.id.latitude)).getText().toString());
-                    Double longitude= Double.parseDouble(((EditText) findViewById(R.id.longitude)).getText().toString());
-                    DataAndMethods.getMap(latitude, longitude);
-                } catch (NumberFormatException e){
-                    speaker("Invalid coordinates");
+                    DataAndMethods.speaker(DataAndMethods.getFile(++fileSelected));
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
                 } catch (JSONException e) {
                     throw new RuntimeException(e);
                 }
-                //startActivity(new Intent(getApplicationContext(), BasicPhotoMapRenderer.class));
+                return true;
+            case "DOWN":
+                Log.d("KEY EVENT", event.toString());
+                try {
+                    DataAndMethods.speaker(DataAndMethods.getFile(--fileSelected));
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                } catch (JSONException e) {
+                    throw new RuntimeException(e);
+                }
+                return true;
+            case "OK":
+                if (update.getValue())
+                    {startActivity(new Intent(getApplicationContext(), BasicPhotoMapRenderer.class));}
+                else
+                    pingsPlayer(R.raw.image_error);
                 return false;
             case "BACK":
                 finish();
@@ -115,14 +108,22 @@ public class MapSelector extends AppCompatActivity implements MediaPlayer.OnComp
 
     @Override
     protected void onResume() {
-        Log.d("ACTIVITY", "MapSelector Resumed");
-        DataAndMethods.speaker("Map selector");
-        image= null;
+        Log.d("ACTIVITY", "PhotoSelector Resumed");
+        DataAndMethods.speaker("Photo selector");
+        try {
+            // might want to read file name here
+            DataAndMethods.getFile(++DataAndMethods.fileSelected);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        } catch (JSONException e) {
+            throw new RuntimeException(e);
+        }
         super.onResume();
     }
     @Override
     protected void onPause() {
-        Log.d("ACTIVITY", "MapSelector Paused");
+        Log.d("ACTIVITY", "PhotoSelector Paused");
         super.onPause();
     }
+
 }
