+++ app/src/main/res/layout/activity_basic_photo_renderer.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BasicPhotoMapRenderer">
+    tools:context=".renderers.BasicPhotoMapRenderer">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
