+++ app/src/main/res/layout/activity_main.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/activity_main_layout"
-    tools:context="ca.mcgill.a11y.image.Exploration">
+    tools:context="ca.mcgill.a11y.image.renderers.Exploration">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
