+++ app/src/main/res/layout/activity_mode_selector.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ModeSelector">
+    tools:context=".selectors.ModeSelector">
 
     <TableLayout
         android:layout_width="409dp"
