+++ app/src/main/res/layout/activity_photo_selector.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PhotoSelector">
+    tools:context=".selectors.PhotoSelector">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
